pyproject.toml
[tool.poetry]
name = "rag-bot"
version = "0.1.0"
description = ""
authors = ["Mikhail Zuev <mikhailneue@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
aiogram = "^3.24.0"
python-dotenv = "^1.2.1"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
pgvector = "^0.4.2"
pinecone = "^8.0.0"
httpx = "^0.28.1"
chromadb = "^1.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bot = "rag_bot.main:run"
